update tooling to rollup, move example, code cleanup
