+++ example/package.json
@@ -1,0 +1,21 @@
+{
+  "scripts": {
+    "start": "snowpack dev",
+    "build": "snowpack build"
+  },
+  "dependencies": {
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "@keboola/flow-builder": "../"
+  },
+  "devDependencies": {
+    "@snowpack/plugin-dotenv": "^2.1.0",
+    "@snowpack/plugin-react-refresh": "^2.5.0",
+    "@snowpack/plugin-typescript": "^1.2.1",
+    "@types/react": "^16.14.0",
+    "@types/react-dom": "^16.9.14",
+    "@types/snowpack-env": "^2.3.3",
+    "snowpack": "^3.3.7",
+    "typescript": "^4.2.4"
+  }
+}
