+++ example/public/favicon.ico
