+++ example/snowpack.config.mjs
