+++ example/src/App.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { Graph, Group, Node, Input, Output } from "./graph"
+import { Graph, Group, Node, Input, Output } from "@keboola/flow-builder";
 
 const App = () => {
   return (
     <div style={{
       position: "relative",
       left: "20px",
-      width: "100%",
-      height: "100%"
+      width: "50%",
+      height: "50%"
     }}>
       <style>
         {`
@@ -17,24 +17,24 @@ const App = () => {
           text-align: center;
           line-height: 3em;
         }
-        .flow-graph--node {
+        .flow-builder--node {
           border: 1px solid rgb(223, 225, 229);
           border-radius: 8px;
           box-shadow: 2px 2px 2px 1px rgba(245, 245, 245, 100);
         }
-        .flow-graph--group {
+        .flow-builder--group {
           border: 1px solid rgb(223, 225, 229);
           border-radius: 8px;
           box-shadow: 2px 2px 4px 2px rgba(245, 245, 245, 200);
         }
-        .flow-graph--group .flow-graph--node {
+        .flow-builder--group .flow-builder--node {
           border: none;
           box-shadow: none;
         }
-        .flow-graph--io-port {
+        .flow-builder--io-port {
           
         }
-        .flow-graph--io-port[data-name$="success"] {
+        .flow-builder--io-port[data-name$="success"] {
           background-color: green;
         }
         `}
@@ -43,15 +43,15 @@ const App = () => {
         "a.success->b+c+d.run0",
         "f.success->b+c+d.run1",
       ]}>
-        <Node align-x name="a" position={[50, 10]}>
+        <Node name="a" position={[50, 50]}>
           <div className="content">Test</div>
           <Output name="success" />
         </Node>
-        <Node align-x name="f" position={[250, 10]}>
+        <Node name="f" position={[250, 50]}>
           <div className="content">Test</div>
           <Output name="success" />
         </Node>
-        <Group align-x name="b+c+d" position={[150, 100]}>
+        <Group name="b+c+d" position={[150, 150]}>
           <Input name="run0" />
           <Input name="run1" />
           <Node name="b">
