+++ src/Graph.css
@@ -1,4 +1,4 @@
-.flow-graph {
+.flow-builder {
   display: block;
   position: relative;
   box-sizing: content-box;
@@ -6,52 +6,50 @@
   height: 100%;
 }
 
-.flow-graph svg {
+.flow-builder svg {
   width: 100%;
   height: 100%;
   pointer-events: none;
 }
 
-.flow-graph--node {
+.flow-builder svg path {
+  stroke: black;
+  fill: transparent;
+}
+
+.flow-builder--node {
   position: absolute;
   display: inline-flex;
   flex-direction: row;
   background-color: transparent;
-}
-
-.flow-graph--align-x {
-  transform: translate(-50%, 0);
-}
-.flow-graph--align-y {
-  transform: translate(0, -50%);
-}
-.flow-graph--align-x.flow-graph--align-y {
   transform: translate(-50%, -50%);
 }
 
-.flow-graph--content {
+.flow-builder--content {
   width: 100%;
   height: 100%;
   display: inline-flex;
   flex-direction: row;
 }
 
-.flow-graph--group {
+.flow-builder--group {
   position: absolute;
   display: inline-flex;
   flex-direction: row;
   background-color: transparent;
+  transform: translate(-50%, -50%);
 }
 
-.flow-graph--group .flow-graph--node .flow-graph--io {
+.flow-builder--group .flow-builder--node {
+  position: static;
+  transform: none;
+}
+
+.flow-builder--group .flow-builder--node .flow-builder--io {
   display: none;
 }
 
-.flow-graph--group .flow-graph--node {
-  position: static;
-}
-
-.flow-graph--io {
+.flow-builder--io {
   position: absolute;
   display: flex;
   justify-content: space-evenly;
@@ -63,17 +61,17 @@
   background-color: transparent;
 }
 
-.flow-graph--io.flow-graph--io-top {
+.flow-builder--io.flow-builder--io-top {
   top: -8px;
   bottom: auto;
 }
 
-.flow-graph--io.flow-graph--io-bottom {
+.flow-builder--io.flow-builder--io-bottom {
   top: auto;
   bottom: -8px;
 }
 
-.flow-graph--io-port {
+.flow-builder--io-port {
   position: static;
   margin: 0;
   width: 16px;
@@ -81,9 +79,4 @@
   border-radius: 100%;
   background-color: red;
   pointer-events: all;
-}
-
-.flow-graph--svg path {
-  stroke: black;
-  fill: transparent;
 }
