+++ src/context.ts
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const GraphContext = createContext<{ parent: string | null }>({ parent: null });
