--- src/graph/context.ts
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const GraphContext = createContext<string | null>(null);
