--- src/graph/index.tsx
@@ -1,5 +1,0 @@
-
-export * from "./Io";
-export * from "./Node";
-export * from "./Group";
-export * from "./Graph";
