+++ src/index.tsx
@@ -1,17 +1,8 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-import "./index.css";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+import "./Graph.css";
+import { Input, Output } from "./Io";
+import { Node } from "./Node";
+import { Group } from "./Group";
+import { Graph } from "./Graph";
 
-// Hot Module Replacement (HMR) - Remove this snippet to remove HMR.
-// Learn more: https://snowpack.dev/concepts/hot-module-replacement
-if (import.meta.hot) {
-  import.meta.hot.accept();
-}
+export { Input, Output, Group, Node, Graph };
